requirements.txt
streamlit==1.32.2
torch>=2.0.0
torchvision>=0.15.0
transformers==4.41.2
pillow>=9.5.0
numpy<2.0
accelerate>=0.27.0